Better estimation of pool max block size distribution
Instead of using blocks where mempool size is largest, use blocks
with the largest mempool size delta against previous block.

The reason for this is that we want to be block-size limited as
often as possible. Previously, mempool size could be huge but txs
largely have fee rate less than the minfeerate of some miners.